renamed main script
